feat: adds login client component and refactors login page
