+++ apps/web/src/lib/auth/index.ts
@@ -1,18 +1,28 @@
 import { authAPI } from '../api';
 import { getAuthFromRequest } from './middleware';
 
-export async function getAuthUser() {
+export async function getAuthFromCookie() {
     const auth = await getAuthFromRequest();
     if (!auth.isAuthenticated) {
+        return null;
+    }
+
+    if (auth.isExpired) {
+        console.error('Token expired, NEED TO REFRESH (TODO)');
         return null;
     }
 
     return auth.user || null;
 }
 
-export async function getAuthUserFromAPI() {
+export async function getAuthFromAPI() {
     const auth = await getAuthFromRequest();
     if (!auth.isAuthenticated) {
+        return null;
+    }
+
+    if (auth.isExpired) {
+        console.error('Token expired, NEED TO REFRESH (TODO)');
         return null;
     }
